+++ server/index.js
@@ -12,7 +12,7 @@ const auth = require("./routes/auth");
 const api = require("./routes/api/index");
 
 const router = require("./routes");
-const { isAuthenticated } = require("./controllers/auth");
+const { isAuthenticated } = require("./controllers/adminController");
 
 const webpackDevMiddleware = require("webpack-dev-middleware");
 const webpackConfig = require("../webpack.dev.js");
